// src/components/IndividualRiskAssesment/types.ts
interface RiskSuccess {
  success: true;
  data: {
    individualRiskAssessment: {
      uuid: string;
      id: number;
    };
  };
}

interface RiskError {
  success: false;
  data: Record<string, string>;
  message?: string;
  status?: number;
  alert?: boolean;
}

export type RiskResponse = RiskSuccess | RiskError;